backend/requirements.txt
pytest
pytest-xdist 	# if you're running tests on GH Codespaces
StackAPI
PyGithub
GitPython
huggingface-hub
transformers
sentence-transformers
python-dotenv
pydantic
pydantic_settings
<<<<<<< Updated upstream
=======
passlib
>>>>>>> Stashed changes
scikit-learn
numpy 
torch
python-jwt
jwt
python-jose
faiss-cpu
cryptography
fastapi
<<<<<<< Updated upstream
passlib
=======
uvicorn
httpx
>>>>>>> Stashed changes
